@@ -1,6 +1,3 @@
-[bdist_wheel]
-universal = 1
-
 [metadata]
 name = read_version
 #version = # Set in setup.py
@@ -45,7 +42,7 @@ project_urls =
 py_modules = read_version
 package_dir =
     =src
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, <4
+python_requires = ~=3.6
 install_requires =
     setuptools
 
